requirements.txt
requests
tda-api
selenium
datetime
requests
matplotlib
pytz
schedule
numpy